pyproject.toml
[tool.poetry]
name = "registration"
version = "0.1.0"
description = ""
authors = ["Name Surname <email@domain.com>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
Flask = "^2.2.2"
psycopg2-binary = "^2.9.4"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
